add dumb-init entry point (#109)
this allows the container instance to handle the SIGTERM correctly and terminate the application.

we are experiencing timeout while stoping the container this fix should handle this situations, without waiting for podman or docker to kill the process.

https://issues.redhat.com/browse/AAP-55964

Signed-off-by: Djebran Lezzoum <ldjebran@gmail.com>
(cherry picked from commit a82bf072eb842103f96e63711bc8917c8744deb3)